wallet withdrawal is done wallet balance enquiry input amount bug fixed

+++ SDK/m90app/xpos-sdk-demo/src/vas/airtime.h
@@ -36,7 +36,7 @@ protected:
 
     std::string phoneNumber;
     PaymentMethod payMethod;
-    unsigned long amount;
+    long long amount;
 
     CardPurchase cardPurchase;
