+++ SDK/m90app/xpos-sdk-demo/src/vas/paytv.h
@@ -30,7 +30,7 @@ protected:
     std::string phoneNumber;
     std::string iuc;
     PaymentMethod payMethod;
-    unsigned long amount;
+    long long amount;
 
     CardPurchase cardPurchase;
